move yaml v3 writing to `_io` module
