+++ src/mammos_entity/_io/__init__.py
@@ -8,7 +8,7 @@ from ._hdf5_v1 import _from_hdf5_v1
 from ._hdf5_v2 import _from_hdf5_v2
 from ._yaml_v1 import _from_yaml_v1
 from ._yaml_v2 import _from_yaml_v2
-from ._yaml_v3 import _from_yaml_v3
+from ._yaml_v3 import _from_yaml_v3, _to_yaml_v3
 
 __all__ = [
     "_from_csv_v1",
@@ -21,4 +21,5 @@ __all__ = [
     "_from_hdf5_v1",
     "_from_hdf5_v2",
     "_to_csv_v4",
+    "_to_yaml_v3",
 ]
